# chatty/delivery/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Chat Server API",
        "contact": {
            "email": "gusiatnikovanatalia@gmail.com"
        },
        "license": {
            "name": "free-to-use-license"
        },
        "version": "1.0"
    },
    "paths": {
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health check"
                ],
                "summary": "Return 200 and empty json if service is OK",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User operations"
                ],
                "summary": "Login user in system by checking the specified password",
                "parameters": [
                    {
                        "description": "Login and password for user",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LoginRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserResponseBody"
                        }
                    },
                    "400": {
                        "description": "Request body is incorrect or data validation have failed"
                    },
                    "404": {
                        "description": "User with the specified login is not exists"
                    },
                    "415": {
                        "description": "Content-Type application/json is missing"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User operations"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "Login, Password, Email, Phone number for user",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RegisterRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/UserResponseBody"
                        }
                    },
                    "400": {
                        "description": "Request body is incorrect or data validation have failed"
                    },
                    "409": {
                        "description": "User with the specified login | email | phone number is already exists"
                    },
                    "415": {
                        "description": "Content-Type application/json is missing"
                    }
                }
            }
        },
        "/user/{login}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User operations"
                ],
                "summary": "Return information about user based on login param",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's login",
                        "name": "login",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserResponseBody"
                        }
                    },
                    "400": {
                        "description": "Data validation have failed"
                    },
                    "404": {
                        "description": "User with the specified login is not exists"
                    }
                }
            }
        },
        "/whoami": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User operations"
                ],
                "summary": "Return information about user based on jwt token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserResponseBody"
                        }
                    },
                    "404": {
                        "description": "User is not found"
                    }
                }
            }
        }
    },
    "definitions": {
        "LoginRequestBody": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string",
                    "example": "testUser123"
                },
                "password": {
                    "type": "string",
                    "example": "q123!@#Q"
                }
            }
        },
        "RegisterRequestBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "example@gmail.com"
                },
                "login": {
                    "type": "string",
                    "example": "testUser123"
                },
                "password": {
                    "type": "string",
                    "example": "q123!@#Q"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+71234567890"
                }
            }
        },
        "UserResponseBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "example@gmail.com"
                },
                "id": {
                    "type": "string",
                    "example": "018496f4-77d7-0ef1-c2d2-f2b09e7b3fb1"
                },
                "login": {
                    "type": "string",
                    "example": "testUser123"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+71234567890"
                }
            }
        }
    }
}